+++ onediff_comfy_nodes/__init__.py
@@ -7,7 +7,6 @@ from ._nodes import (
     VaeSpeedup,
     VaeGraphLoader,
     VaeGraphSaver,
-    ControlNetSpeedup,
     SVDSpeedup,
     ModuleDeepCacheSpeedup,
     OneDiffCheckpointLoaderSimple,
@@ -25,7 +24,6 @@ NODE_CLASS_MAPPINGS = {
     "VaeSpeedup": VaeSpeedup,
     "VaeGraphSaver": VaeGraphSaver,
     "VaeGraphLoader": VaeGraphLoader,
-    "ControlNetSpeedup": ControlNetSpeedup,
     "SVDSpeedup": SVDSpeedup,
     "ModuleDeepCacheSpeedup": ModuleDeepCacheSpeedup,
     "OneDiffCheckpointLoaderSimple": OneDiffCheckpointLoaderSimple,
@@ -41,7 +39,6 @@ NODE_DISPLAY_NAME_MAPPINGS = {
     "VaeSpeedup": "VAE Speedup",
     "VaeGraphLoader": "VAE Graph Loader",
     "VaeGraphSaver": "VAE Graph Saver",
-    "ControlNetSpeedup": "ControlNet Speedup",
     "SVDSpeedup": "SVD Speedup",
     "ModuleDeepCacheSpeedup": "Model DeepCache Speedup",
     "OneDiffCheckpointLoaderSimple": "Load Checkpoint - OneDiff",
